+++ NEWS.md
@@ -1,7 +1,7 @@
 PENDING: Dataiter 1.1
 =====================
 
-* `DataFrame.pivot_longer`: TODO
+* `DataFrame.pivot_longer`: New method
 * `DataFrame.pivot_wider`: New method
 
 2025-02-07: Dataiter 1.0